implement object static classes
