+++ dotnet/RarelySimple.AvatarScriptLink.Net.Tests/Decorators/FormObjectDecoratorTests.cs
@@ -1,6 +1,7 @@
 ﻿using RarelySimple.AvatarScriptLink.Net.Decorators;
 using RarelySimple.AvatarScriptLink.Net.Exceptions;
 using RarelySimple.AvatarScriptLink.Objects;
+using static RarelySimple.AvatarScriptLink.Objects.RowObject;
 
 namespace RarelySimple.AvatarScriptLink.Net.Tests.Decorators;
 
@@ -579,7 +580,7 @@ public class FormObjectDecoratorTests
         var rowObject2 = new RowObject()
         {
             RowId = "456||2",
-            RowAction = RowAction.Delete
+            RowAction = RowActions.Delete
         };
         var formObject = new FormObject()
         {
@@ -603,7 +604,7 @@ public class FormObjectDecoratorTests
         var rowObject2 = new RowObject()
         {
             RowId = "456||2",
-            RowAction = RowAction.Delete
+            RowAction = RowActions.Delete
         };
         var formObject = new FormObject()
         {
