+++ dotnet/RarelySimple.AvatarScriptLink.Objects/RowObject.cs
@@ -110,5 +110,28 @@ namespace RarelySimple.AvatarScriptLink.Objects
         }
 
         #endregion
+
+        /// <summary>
+        /// Represents a valid AvatarScriptLink <see cref="RowAction"/>.
+        /// </summary>
+        public static class RowActions
+        {
+            /// <summary>
+            /// Instructs myAvatar to add this <see cref="RowObject"/>.
+            /// </summary>
+            public const string Add = "ADD";
+            /// <summary>
+            /// Instructs myAvatar to delete this <see cref="RowObject"/>.
+            /// </summary>
+            public const string Delete = "DELETE";
+            /// <summary>
+            /// Instructs myAvatar to edit this <see cref="RowObject"/>.
+            /// </summary>
+            public const string Edit = "EDIT";
+            /// <summary>
+            /// Instructs myAvatar to do nothing with this <see cref="RowObject"/>.
+            /// </summary>
+            public const string None = "";
+        }
     }
 }
